src/requirements.txt
django
pillow
djangorestframework
django-stubs
djangorestframework-stubs